fix(app & components): organize files and folders structure
